Move file sync to its own package
